Fixes various errors within the CSharp tests

+++ csharp/Answers/Question1.cs
@@ -8,7 +8,7 @@ namespace C_Sharp_Challenge_Skeleton.Answers
 {
     public class Question1
     {
-        public static int Answer(double initialLevelOfDebt, double interestPercentage, double repaymentPercentage)
+        public static int Answer(decimal initialLevelOfDebt, decimal interestPercentage, decimal repaymentPercentage)
         {
             //TODO: Please work out the solution;
             return -1;
